requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
git+https://github.com/pyrogram/pyrogram@master
youtube-search-python
yt-dlp
speedtest-cli
youtube-dl
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
pillow
motor
psutil
future
wget
lyricsgenius
yt-dlp==2021.11.10.1
skem
